+++ src/main/kotlin/com/github/mrdolch/plantarchintellijplugin/diagram/view/DiagramView.kt
@@ -49,7 +49,7 @@ object DiagramView {
         application.runWriteAction { VfsUtil.saveText(it, plantuml) }
         fileEditorManager.openFile(it, true, true)
         EditorRegistry.getEditor(it)!!
-          .updateFields(plantuml)
+          .updateFields(it, plantuml)
       }
     }
   }